+++ packages/at_chops/lib/src/algorithm/algo_type.dart
@@ -3,7 +3,7 @@ import 'package:at_commons/at_commons.dart';
 
 // ignore: constant_identifier_names
 enum SigningAlgoType {
-  ecc_secp256r1,
+  eccSecp256r1,
   rsa2048,
   rsa4096,
   ed25519,