+++ packages/at_chops/lib/src/algorithm/signing/ed25519.dart
@@ -1,13 +1,12 @@
 import 'dart:async';
 import 'dart:typed_data';
 
-import 'package:at_chops/src/algorithm/at_algorithm.dart';
 import 'package:at_commons/at_commons.dart';
 import 'package:better_cryptography/better_cryptography.dart';
 
 /// Data signing and verification for Ed25519 - elliptic curve algorithm
-/// Keypair for the algorithm has to generated using [AtChopsUtil.generateEd25519KeyPair()]
-class Ed25519SigningAlgo implements AtSigningAlgorithm {
+/// Set the keypair via [ed25519KeyPair] before signing.
+class Ed25519SigningAlgo {
   final _algorithm = Ed25519();
   SimpleKeyPair? _ed25519KeyPair;
 
@@ -17,7 +16,6 @@ class Ed25519SigningAlgo implements AtSigningAlgorithm {
 
   Ed25519SigningAlgo();
 
-  @override
   Future<Uint8List> sign(Uint8List data) async {
     if (_ed25519KeyPair == null) {
       throw AtSigningException(
@@ -27,7 +25,6 @@ class Ed25519SigningAlgo implements AtSigningAlgorithm {
     return Uint8List.fromList(signature.bytes);
   }
 
-  @override
   Future<bool> verify(Uint8List signedData, Uint8List signature,
       {String? publicKey}) async {
     if (publicKey == null) {